# editor/server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "editor API",
    "version": "0.1.0",
    "description": "Jinja2 テンプレート GUI 編集アプリのバックエンド API。\n\n`@editor/shared` の Repository 契約(Template / Auth / User / History / Part)を REST として設計したもの。一部は Phase 2 の設計(ハンドラ未実装)。"
  },
  "servers": [
    {
      "url": "/api",
      "description": "同一オリジン (server が SPA も配信)"
    }
  ],
...
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/templates/options": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "カスケードドロップダウンの候補を取得",
        "operationId": "getDropdownOptions",
        "parameters": [
          {
            "in": "query",
            "name": "companyCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fundCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "baseDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "editionType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string",
              "enum": [
                "edit",
                "published"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "各属性の候補",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DropdownOptions"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/templates": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "テンプレート一覧(属性で絞り込み)",
        "operationId": "listTemplates",
        "parameters": [
          {
            "in": "query",
            "name": "companyCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fundCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "baseDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "editionType",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "テンプレート meta の配列",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateMeta"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/companies": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "作成タブの委託会社一覧(Rep1 のファンド属性)",
        "operationId": "listCompanies",
        "responses": {
          "200": {
            "description": "委託会社(略称 = ファイル名の会社コード)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyOption"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/funds": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "作成タブのファンド一覧(Rep1 の委託会社コードで引く)",
        "operationId": "listFunds",
        "parameters": [
          {
            "in": "query",
            "name": "rep1CompanyCode",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ファンドの一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FundOption"
                  }
                }
              }
            }
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/creatable": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "作成タブの作成可否(作成済み・シリーズのコピー元候補)",
        "operationId": "getCreatableInfo",
        "parameters": [
          {
            "in": "query",
            "name": "companyCode",
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true
          },
          {
            "in": "query",
            "name": "rep1CompanyCode",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32
            },
            "required": true
          },
          {
            "in": "query",
            "name": "fundCode",
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true
          },
          {
            "in": "query",
            "name": "editionType",
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "作成済みとシリーズのコピー元候補",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatableInfo"
                }
              }
            }
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{id}": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "テンプレート(meta + html + css)を取得",
        "operationId": "getTemplate",
        "description": "確定テンプレを先に見る。無い場合だけ生成直後の未確定(pending)実体を `status: \"draft\"` で返す。順序が逆だと pending を書ける者が確定テンプレの表示内容を差し替えられるため、確定優先は契約である。",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "テンプレート",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Template"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "404": {
            "description": "対象が存在しない (kind=not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
...
    "/templates/{id}/draft": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "常時オートセーブの下書きを取得",
        "operationId": "getDraft",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "下書き(無ければ null)",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TemplateDraft"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "templates"
        ],
        "summary": "下書きを保存(オートセーブ)",
        "operationId": "saveDraft",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDraftRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "保存完了"
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "templates"
        ],
        "summary": "下書きを破棄(確定保存せずメニューへ戻る時)",
        "operationId": "discardDraft",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "破棄完了(冪等)"
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
...
    "/funds/{fundCode}/sample-data": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "プレビュー用サンプルデータ(fundCode ごと)",
        "operationId": "getSampleData",
        "parameters": [
          {
            "in": "path",
            "name": "fundCode",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "サンプルデータ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleData"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "404": {
            "description": "対象が存在しない (kind=not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SaveDraftRequest": {
        "type": "object",
        "properties": {
          "templateId": {
            "$ref": "#/components/schemas/TemplateId"
          },
          "html": {
            "type": "string"
          },
          "css": {
            "type": "string"
          }
        },
        "required": [
          "templateId",
          "html",
          "css"
        ]
      },
      "TemplateId": {
        "type": "string",
        "example": "AM01_510037_20240710_交付版",
        "description": "ファイル名規約(拡張子なし)。値入り HTML は `<会社コード>_<ファンドコード>_<基準日>_<版種>`、テンプレートは `<会社コード>_<ファンドコード>_<版種>`。パス区切り・`..`・制御文字・末尾のドット/空白を含まない単一のファイル名セグメントに限る(判定は `isValidAnyTemplateId`)。"
      },
...
      "AppError": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/AppErrorKind"
          },
          "message": {
            "type": "string",
            "description": "ユーザ向け(JP)。常に表示して安全"
          },
          "code": {
            "description": "機械可読コード 例: 'USER_DISABLED'",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "message"
        ],
        "additionalProperties": false
      },
      "AppErrorKind": {
        "type": "string",
        "enum": [
          "not_found",
          "validation",
          "unauthorized",
          "forbidden",
          "conflict",
          "network",
          "unexpected"
        ]
      },
      "DropdownOptions": {
        "type": "object",
        "properties": {
          "companyCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fundCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "baseDates": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "editionTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "companyCodes",
          "fundCodes",
          "baseDates",
          "editionTypes"
        ],
        "additionalProperties": false
      },
      "TemplateMeta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ファイル名(拡張子なし)由来の安定 ID"
          },
          "attributes": {
            "$ref": "#/components/schemas/TemplateAttributes"
          },
          "fileName": {
            "type": "string",
            "example": "AM01_510037_20240710_kr.html"
          },
          "status": {
            "$ref": "#/components/schemas/TemplateStatus"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "最終確定保存の ISO タイムスタンプ"
          },
          "updatedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "attributes",
          "fileName",
          "status",
          "updatedAt",
          "updatedBy"
        ],
        "additionalProperties": false
      },
      "TemplateAttributes": {
        "type": "object",
        "properties": {
          "companyCode": {
            "type": "string",
            "description": "委託会社(ファイル名の会社コード = Rep1 の委託会社略称)"
          },
          "fundCode": {
            "type": "string",
            "description": "ファンドコード"
          },
          "baseDate": {
            "description": "基準日(yyyymmdd または yyyy-mm-dd)。値入り HTML(filled/)だけが持ち、テンプレート(templates/)は持たない",
            "example": "2024-05-17",
            "type": "string"
          },
          "editionType": {
            "type": "string",
            "description": "版種"
          }
        },
        "required": [
          "companyCode",
          "fundCode",
          "editionType"
        ],
        "additionalProperties": false
      },
      "TemplateStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published"
        ]
      },
      "CompanyOption": {
        "type": "object",
        "properties": {
          "companyCode": {
            "type": "string",
            "description": "ファイル名の会社コード(Rep1 の委託会社略称)"
          },
          "companyName": {
            "type": "string",
            "description": "委託会社名"
          },
          "rep1CompanyCode": {
            "type": "string",
            "description": "Rep1 の委託会社コード(ファンドを引くときに使う)"
          }
        },
        "required": [
          "companyCode",
          "companyName",
          "rep1CompanyCode"
        ],
        "additionalProperties": false
      },
      "FundOption": {
        "type": "object",
        "properties": {
          "fundCode": {
            "type": "string"
          },
          "fundName": {
            "type": "string"
          }
        },
        "required": [
          "fundCode",
          "fundName"
        ],
        "additionalProperties": false
      },
      "CreatableInfo": {
        "type": "object",
        "properties": {
          "created": {
            "type": "boolean",
            "description": "選んだ会社・ファンド・版種のテンプレートが templates/ にあるか(会社_ファンド_版種.html。大文字小文字は区別しない)"
          },
          "templateId": {
            "description": "作成済みのときのテンプレートの id(templates/ のファイルの綴りのまま)。作成タブの「既存のテンプレートを開く」で開く",
            "type": "string"
          },
          "inProgressId": {
            "description": "作成済みでなく、同じ id の下書きか pending/ があるときの id。作成タブの「作成中のテンプレートを開く」で開く。作り直すときは確認のうえ GenerateRequest.replaceExisting を付ける",
            "type": "string"
          },
          "seriesFunds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeriesFundOption"
            },
            "description": "同じシリーズの他のファンド(シリーズから作成のコピー元候補)"
          }
        },
        "required": [
          "created",
          "seriesFunds"
        ],
        "additionalProperties": false
      },
      "SeriesFundOption": {
        "type": "object",
        "properties": {
          "fundCode": {
            "type": "string"
          },
          "fundName": {
            "type": "string"
          },
          "hasTemplate": {
            "type": "boolean",
            "description": "コピー元のテンプレート(templates/<会社>_<ファンド>_<版種>.html。基準日なし)があるか"
          }
        },
        "required": [
          "fundCode",
          "fundName",
          "hasTemplate"
        ],
        "additionalProperties": false
      },
      "Template": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/TemplateMeta"
          },
          "html": {
            "type": "string",
            "description": "Jinja2 生 HTML(タグ保持)"
          },
          "css": {
            "type": "string",
            "description": "テンプレ単位の CSS(css/<会社>_<ファンド>_<版種>.css。基準日違いの文書で共有)"
          },
          "filled": {
            "type": "string",
            "description": "エディタキャンバス用に事前描画した filled HTML(Jinja 値を差し込みつつ元ソースを保持)。静的な fill が無ければ空で、エディタは都度描画にフォールバックする"
          },
          "cssMissing": {
            "description": "確定版の CSS ファイル(css/<会社>_<ファンド>_<版種>.css)が見つからないとき true。あれば付けない。生成直後(pending)は pending の CSS ファイルが無いとき true",
            "type": "boolean"
          }
        },
        "required": [
          "meta",
          "html",
          "css",
          "filled"
        ],
        "additionalProperties": false
      },
...
      "TemplateDraft": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "string"
          },
          "html": {
            "type": "string"
          },
          "css": {
            "type": "string"
          },
          "savedAt": {
            "type": "string"
          },
          "savedBy": {
            "type": "string"
          }
        },
        "required": [
          "templateId",
          "html",
          "css",
          "savedAt",
          "savedBy"
        ],
        "additionalProperties": false
      },
...
      "SampleData": {
        "type": "object",
        "propertyNames": {
          "type": "string"
        },
        "additionalProperties": {},
        "description": "プレビュー用サンプルデータ(任意の JSON)"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "editor.sid",
        "description": "ログイン時に発行されるセッション cookie。"
      }
    }
  }
}